+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,7 @@ upcoming
 ++++++
 * 'az containerapp create': Fix Role assignment error when the default Azure Container Registry could not be found
 * Upgrade api-version to 2024-10-02-preview
+* 'az containerapp create/update': `--yaml` support property pollingInterval and cooldownPeriod
 
 1.0.0b4
 ++++++